Removed root_url setting of grafana
With the root_url setting, it will result in a error page showing "If you're seeing this Grafana has failed to load its application files" when visit the Cost Report  refering to https://github.com/gocrane/crane

AFAIK，by referring with https://grafana.com/docs/grafana/latest/administration/configuration/,   the root_url setting is usually used with grafana deployment behind reverse proxy or using Google or GitHub OAuth authentication (for the callback URL to be correct).  It should be removed if we just run a demo.